Refactor boolean type definition for C++/-std=gnu23 compatibility
Changed the definition of 'boolean' type for C++ & -std=gnu23 compatibility, and included base <stdbool.h> header.

This is to help bring it more inline with other source ports (Chocolate Doom mainly) and keep compatibility with new toolchains I think. I didn't come up with this fix (https://blog.svgames.pl/article/the-little-bool-of-doom) but it's helped me work on a port of doomgeneric to PlayStation 2

How easy would it be to get audio working? I think only SDL 1.2 works well on that platform so I'll be using GSKit instead